SPF: Change navigation history/request order so the Referer header is usually correct.
Now, standard navigation in response to a click or JS navigate call will have the correct referrer sent.  This is done by sending the XHR for the next page before adding the new history entry.  However, back/forward navigation in response to browser history events will still be wrong because the browser changes the URL before the XHR is sent.  In both cases, an additional X-SPF-Referer header will be sent with the correct value.

Replace the "navigate-started-callback" and "navigate-history-callback" notifications with a single unified "navigate-requested-callback" notification that is published for either link clicks or history events.  Remove unused "load-received-callback" and "load-processed-callback" notifications, which are redundant with passed callback functions.

Also, minor fixes to the demo app to properly display the referer and to mark external scripts as loaded.


[git-p4: change = 47724795]