SPF: Add XHR processing support for chunked transfers and add Response processing support for multipart responses. Used together, an SPF request can be processed in stages as it arrives, similar to standard HTML rendering.
This CL introduces multipart responses and ensures that they can be used
interchangeably.  It also starts splitting navigation-related methods
into functional namespaces and unit tests.

The unchanged transfer format for a single (traditional) response is
an object serialized in JSON:
  '{ "css": "...", "html": {...}, ..., "js": "..." }'

The unchanged object for a single response after parsing this format is:
  {
    'css': '...',
    'html': { ... },
    ...,
    'js': '...',
    'timing': { ... }
  }

The new transfer format for a multipart response is an array of objects
serialized in JSON, where '[\r\n', ',\r\n', and ']\r\n' delimiters
are used to facilitate on-the-fly parsing:
  '[\r\n
   {"css": "..."},\r\n
   {"html": {...}},\r\n
   ...,\r\n
   {"js": "..."}
   ]\r\n'

The new object for a multipart response after parsing this format is:
  {
    'type': 'multipart'
    'parts: [
      {'css': '...'},
      {'html': { ... }},
      ...
      {'js': '...'}
    ],
    'timing': { ... }
  }



The additional code to handle the multipart responses has increased the binary size:
File      Before After Change  Percent
spf.js     14222 16098  +1876  +13.19%
spf.js.gz   5160  5835   +675  +13.08%


[git-p4: change = 53682233]