This fixes SPF prefetching errors with IE and Firefox.
The IE bug was a simple typo which created a separate script tag, but forgot to add it to the page.

The Firefox bug relates to DOM readiness of the iframe. The 0ms timeout is not enough to ensure the iframe has loaded it's DOM. If the DOM is not loaded, Firefox will ignore the request to append a child and the script will never be prefetched. This has been changed to use onload instead.


[git-p4: change = 54961979]