Allow for partial responses during transitions. This requires two changes to spf.
First, SPF now sends an additional request header 'X-SPF-Previous-URL'. This contains the URL of the page which proceeds the loading page. It can be used to make decisions about what parts of the page to reload. It's usually the same as the referer, but differs when history events occur. In those cases the Previous-URL header represents the previously visible page, while the referer represents the original page that lead into the target.

Second, SPF now accepts a cache-suffix in the response. This is appended to the cache key to limit responses which should only be used in certain situations. This is then matched up against the lead in page's URL to determine if cache can be used.

I like the flexibility of the suffix in the response, but I'm not in love with the key generation mechanism, so I'm very open to suggestions here.


[git-p4: change = 61106596]