Move remaining users of history.pushState to History.prototype.
Noticed the exception message was less helpful than
before. This update makes usage consistent and continues to use
the experiment to find 3rd party modifications (this time on
History.prototype).

To precise: if a 3rd party script loads before the history
prototype is secured then we cannot detect it. We assume there
is a race condition and we will detect some modifications after
we already took a copy of the pointers.


[git-p4: change = 66316433]