Scroll elements into view after navigation.
When navigating to a URL with a hash (aka fragment) that references a page
element, scroll the element into view before completing the navigation.  This
means that for a URL such as `/page#element-id`, the `scrollIntoView` function
is called on the referenced element just before the `spfdone` event is fired.

Closes #73.