Clear Performance Resource Timings before each navigate.
Ensure that Resource Timings are not incorrectly retrieved for previous requests
in Chrome where the API is asynchronous, and ensure that the buffer size limit
is not reached.  Clearing is done only before sending new navigate requests to
avoid accidentally removing unrelated Resource Timings during a prefetch or load
request.

As an advanced option, allow the Resource Timings to persist across navigations
if the `advanced-navigate-persist-timing` config is set.  The application must
then be aware of the asynchronous nature of the Chrome API and manage the buffer
size limit accordingly.

Closes #115.