Support configurable persistent parameters.
New advanced config option `advanced-persistent-parameters`. When set, URLs are modified to include the parameters.

Parameters are appended prior to loads and prefetches in addition to navigations. The format of the config should be "abc=def&foo=bar" without a leading "?". The leading delimeter is determined at runtime based on the presence of params on the given URL.

Closes #134