Website: "Pretty" format SPF JSON files to reduce diff sizes.
When updating to website content, it is difficult to see the changes to the
JSON files because the default format places everything on one line.  Use
the "pretty" format to place each key/value on its own line to reduce diff
sizes and make changes more clear.  Since the responses are small, this should
have no measurable impact on site performance.