Website: Adjust scroll-to-top behavior to fix target links.
Previously, the `spfdone` event handler was scrolling to the top of the page
after navigation is complete.  However, if a target link is clicked (one with
a hash), then SPF scrolls to the target just before firing the `spfdone`
event.  To fix this, scroll to the top in the `spfprogress` event handler.