Prevent early scrolling during history navigation (experimental).
When handling a history change event, save the current position and scroll to
that (current) position immediately after handling the event to prevent the
browser from incorrectly scrolling too early.

The browser default behavior is to scroll to the position when `pushState` was
called just after firing a `popState` event.  This is okay only if using
`pushState` to move around a single page or if all content can be rendered
synchronously during the `popState` event handling.  Since navigation content
updates are asynchronous (either from cache or from network), avoid this early
scroll by immediately scrolling back to the current position after the event.
The proper position is then (re)set once the content is updated.

For now, guard this behavior behind the `experimental-scroll-position` config.

Progress on #285.