Alternate approach to prevent early scrolling during navigation.
The current approach to preventing early scrolling during navigation uses the
the current position during a history change event and scrolls to that position
after the event has been handled using `setTimeout`.  This leads to a brief
flash of incorrect position (aka jank) during history navigation because too
much time elapses between the browser's early scroll and the reset call.

Instead, during the history change event, save the current position.  During
the next scroll, if a position has been saved, scroll to it immediately (and
delete the saved position).  This executes the reset call in the same
synchronous operation as the browser's early scroll and eliminates the flash
of incorrect position.

Progress on #285.