Synchronously handle cached responses (experimental)
To avoid the forced 1s delay in WebKit browsers when handling cached responses,
synchronously execute the handler function instead of using a `setTimeout` of 0.
This means that handling a cache response will be synchronous and handling
a network response will be asynchronous.  For evaluation and testing, guard
this behavior behind a `experimental-sync-response-cache` config flag.

Progress on #337.