Synchronously handle only background cached responses (experimental)
To avoid the forced 1s delay in WebKit browsers when handling cached responses
when in the background, synchronously execute the handler function instead of
using a `setTimeout` of 0, but only if `document.hidden` is `true`.  Note that
this still will apply the synchronous behavior to all browsers, but limit the
behavior to tabs that aren't visible.  For evaluation and testing, guard this
behavior behind a `experimental-sync-response-cache-background` config flag.

Progress on #337