Make preconnect requests faster by avoiding DOM changes.
Preconnect requests are done by sending image requests to the target URL.
Before, this was done using `img` tags in an `iframe`, just like prefetching
JS/CSS uses `script` or `link` in an `iframe`.  However, this is unnecessary
overhead that can slow down sending the request.  Instead, use a simple `Image`
object to trigger the request.

Progress on #124.