Release "handle cached responses via deferred execution"
Promote previous `experimental-defer-response-cache` logic to always be enabled.
Remove alternate and conflicting `experimental-sync-response-cache` and
`experimental-sync-response-cache-background` logic.

This avoids the forced 1s delay in WebKit browsers when handling cached
responses in the background by deferring execution of the handler
function using the `spf.async` package, which attempts to use `postMessage`
instead of `setTimeout` for browsers that support it.

Closes #337