Allow custom headers to be sent with requests
In addition to the standard SPF headers (e.g. `X-SPF-Referer`), this change
allows the client to specify custom headers to send with requests.

These headers will be set globally using the `request-headers` config:
```js
spf.init({
  'request-headers': {
    'X-Custom-Header-One': 'Custom Value One',
    'X-Custom-Header-Two': 'Custom Value Two'
  }
})
```
Or, they may be set on a per-request basis for API calls using options:
```js
spf.navigate(url, {
  headers: {
    'X-Custom-Header-One': 'Custom Value One',
    'X-Custom-Header-Two': 'Custom Value Two'
  }
})
```

For now, gate the feature using the `experimental-request-headers` config
and the `experimental_headers` option.

Progress on #390