Create a script to update the version of SPF and generate docs
This change adds a `version.sh` script that updates the version
of SPF and then updates the generated documentation.  Each
update is recorded as a single commit, and when the branch is
merged, the merge commit will be used as the release tag.  This
will mean that tagged releases will included the correct
generated documentation for the version being released.

Progess on #406